+++ openghg/standardise/emissions/__init__.py
@@ -1,2 +1,2 @@
 from ._openghg import parse_openghg
-from ._openghg import parse_intem
+from ._intem import parse_intem
